R04 | T01 | Cadastro de Cliente
- Adiciona components para `register_page`
