+++ lib/app/modules/register/register_page.dart
@@ -2,6 +2,8 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:pscomidas/app/modules/register/register_store.dart';
 import 'package:flutter/material.dart';
+import 'package:pscomidas/app/modules/register/widgets/custom_submit_button.dart';
+import 'package:pscomidas/app/modules/register/widgets/custom_text_field.dart';
 
 class RegisterPage extends StatefulWidget {
   final String title;
@@ -74,118 +76,17 @@ class RegisterPageState extends State<RegisterPage> {
                   ),
                   const SizedBox(height: 20),
                   CustomTextField(
-                    controller: controller,
+                    controller: store.emailController,
                     title: 'E-mail',
                     hint: 'Insira seu email',
                   ),
-                  CustomSubmit(
-                    label: 'Enviar',
-                  )
+                  const SizedBox(height: 10),
+                  const CustomSubmit(label: 'Enviar')
                 ],
               ),
             ),
           ),
         ),
-      ),
-    );
-  }
-}
-
-class CustomTextField extends StatelessWidget {
-  const CustomTextField({
-    Key? key,
-    this.title,
-    this.validator,
-    this.hint,
-    required this.controller,
-  }) : super(key: key);
-
-  final String? title;
-  final String? hint;
-  final String? Function(String?)? validator;
-  final TextEditingController controller;
-
-  TextStyle get titleStyle =>
-      GoogleFonts.getFont('Sen', fontSize: 16.0, fontWeight: FontWeight.bold);
-
-  TextStyle get digitedText => GoogleFonts.getFont('Sen', fontSize: 16.0);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(5.0),
-      child: Column(
-        children: [
-          if (title != null) ...[
-            Row(
-              children: [
-                Expanded(
-                  child: Text(
-                    title!,
-                    style: titleStyle,
-                    textAlign: TextAlign.start,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 5.0),
-          ],
-          TextFormField(
-            style: digitedText,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(
-                borderSide: BorderSide(color: Colors.black26),
-              ),
-              focusedBorder: const OutlineInputBorder(
-                borderSide: BorderSide(color: Colors.red),
-              ),
-              hintText: hint ?? '',
-            ),
-            cursorColor: Colors.red,
-            autovalidateMode: AutovalidateMode.onUserInteraction,
-            controller: controller,
-            validator: validator ??
-                (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'O campo é obrigatório';
-                  }
-                },
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class CustomSubmit extends StatefulWidget {
-  const CustomSubmit({Key? key, required this.label}) : super(key: key);
-  final String label;
-  @override
-  _CustomSubmitState createState() => _CustomSubmitState();
-}
-
-class _CustomSubmitState extends State<CustomSubmit> {
-  TextStyle get digitedText => GoogleFonts.getFont('Sen', fontSize: 22.0);
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(5.0),
-      child: Row(
-        children: [
-          Expanded(
-            child: ElevatedButton(
-              style: ElevatedButton.styleFrom(
-                fixedSize: const Size.fromHeight(50),
-                primary: Colors.red,
-              ),
-              onPressed: () {},
-              child: Text(
-                widget.label,
-                style: digitedText,
-              ),
-            ),
-          )
-        ],
       ),
     );
   }
