+++ lib/app/modules/register/register_store.dart
@@ -1,15 +1,17 @@
+import 'package:flutter/cupertino.dart';
 import 'package:mobx/mobx.dart';
 
 part 'register_store.g.dart';
 
 class RegisterStore = _RegisterStoreBase with _$RegisterStore;
-abstract class _RegisterStoreBase with Store {
 
-  @observable
-  int value = 0;
+abstract class _RegisterStoreBase with Store {
+  TextEditingController nameController = TextEditingController();
+  TextEditingController cpfController = TextEditingController();
+  TextEditingController phoneController = TextEditingController();
+  TextEditingController bornController = TextEditingController();
+  TextEditingController emailController = TextEditingController();
 
   @action
-  void increment() {
-    value++;
-  } 
+  Future<void> register() async {}
 }
